refactor: use composable for OpenCollective stats and drop api route
